FER-147: implementation and testing for persistent local entity with reference to local transient entity;  persistent local entity with reference to remote transient entity; and transient local entity with reference to remote persistent entity

+++ stout/stout-cookbook/stout-cookbook-referencing-domain/src/test/java/org/bitbucket/askllc/fermenter/cookbook/referencing/RemoteReferenceRetrievalIT.java
@@ -50,8 +50,13 @@ public class RemoteReferenceRetrievalIT extends RunTestsOutsideOfArquillianWar {
 
     @After
     public void deleteSimpleDomains() throws Exception {
-        localDomain.delete();
-        referenceMaintenanceDelegate.delete(reference.getId());
+        if (localDomain != null) {
+            localDomain.delete();
+        }
+        if (reference != null) {
+            referenceMaintenanceDelegate.delete(reference.getId());
+        }
+        
         MessageTestUtils.assertNoErrorMessages();
     }
 
@@ -69,7 +74,7 @@ public class RemoteReferenceRetrievalIT extends RunTestsOutsideOfArquillianWar {
         local.setExternalReference(reference);
         localDomain = local.save();
         
-        LocalDomainBO retreivedLocal = LocalDomainBO.findByPrimaryKey(local.getKey());
+        LocalDomainBO retreivedLocal = LocalDomainBO.findByPrimaryKey(localDomain.getKey());
         assertNotNull("Could not find newly minted local domain!", retreivedLocal);
         ValidationReferencedObject retreivedReference = retreivedLocal.getExternalReference();
         assertNotNull("Should have invoked the remote retrieval service!", retreivedReference);
